+++ tools/BuildPackagesTask/Microsoft.Azure.Build.Tasks/CIFilterTask.cs
@@ -70,6 +70,7 @@ namespace Microsoft.WindowsAzure.Build.Tasks
 
         private const string BUILD_PHASE = "build";
         private const string ANALYSIS_BREAKING_CHANGE_PHASE = "breaking-change";
+        private const string ANALYSIS_HELP_EXAMPLE_PHASE = "help-example";
         private const string ANALYSIS_HELP_PHASE = "help";
         private const string ANALYSIS_DEPENDENCY_PHASE = "dependency";
         private const string ANALYSIS_SIGNATURE_PHASE = "signature";
@@ -186,6 +187,7 @@ namespace Microsoft.WindowsAzure.Build.Tasks
                 [BUILD_PHASE] = new HashSet<string>(GetBuildCsprojList(TargetModule, csprojMap).ToList()),
                 [ANALYSIS_BREAKING_CHANGE_PHASE] = new HashSet<string>(GetDependenceModuleList(TargetModule, csprojMap).ToList()),
                 [ANALYSIS_DEPENDENCY_PHASE] = new HashSet<string>(GetDependenceModuleList(TargetModule, csprojMap).ToList()),
+                [ANALYSIS_HELP_EXAMPLE_PHASE] = new HashSet<string>(GetDependenceModuleList(TargetModule, csprojMap).ToList()),
                 [ANALYSIS_HELP_PHASE] = new HashSet<string>(GetDependenceModuleList(TargetModule, csprojMap).ToList()),
                 [ANALYSIS_SIGNATURE_PHASE] = new HashSet<string>(GetDependenceModuleList(TargetModule, csprojMap).ToList()),
                 [TEST_PHASE] = new HashSet<string>(GetTestCsprojList(TargetModule, csprojMap).ToList())
@@ -240,6 +242,7 @@ namespace Microsoft.WindowsAzure.Build.Tasks
                         BUILD_PHASE + ":" + AllModule,
                         ANALYSIS_BREAKING_CHANGE_PHASE + ":" + AllModule,
                         ANALYSIS_DEPENDENCY_PHASE + ":" + AllModule,
+                        ANALYSIS_HELP_EXAMPLE_PHASE + ":" + AllModule,
                         ANALYSIS_HELP_PHASE + ":" + AllModule,
                         ANALYSIS_SIGNATURE_PHASE + ":" + AllModule,
                         TEST_PHASE + ":" + AllModule,
@@ -287,6 +290,7 @@ namespace Microsoft.WindowsAzure.Build.Tasks
                 BUILD_PHASE,
                 ANALYSIS_BREAKING_CHANGE_PHASE,
                 ANALYSIS_DEPENDENCY_PHASE,
+                ANALYSIS_HELP_EXAMPLE_PHASE,
                 ANALYSIS_HELP_PHASE,
                 ANALYSIS_SIGNATURE_PHASE,
                 TEST_PHASE
@@ -337,7 +341,7 @@ namespace Microsoft.WindowsAzure.Build.Tasks
                     }
                     if (csprojSet.Count != 0)
                     {
-                        foreach (string filename in Directory.GetFiles(@"src/Accounts", "*.csproj", SearchOption.AllDirectories).Where(x => !x.Contains("Test")))
+                        foreach (string filename in Directory.GetFiles(@"src\Accounts", "*.csproj", SearchOption.AllDirectories).Where(x => !x.Contains("Test")))
                         {
                             csprojSet.Add(filename);
                         }
@@ -353,7 +357,7 @@ namespace Microsoft.WindowsAzure.Build.Tasks
                     }
                     if (csprojSet.Count != 0)
                     {
-                        csprojSet.Add("tools/TestFx/TestFx.csproj");
+                        csprojSet.Add(@"tools\TestFx\TestFx.csproj");
                     }
                     influencedModuleInfo[phaseName] = csprojSet;
                 }
@@ -397,7 +401,32 @@ namespace Microsoft.WindowsAzure.Build.Tasks
             
             FilterTaskResult.PhaseInfo = influencedModuleInfo;
 
+            if (!Directory.Exists(config.ArtifactPipelineInfoFolder))
+            {
+                Directory.CreateDirectory(config.ArtifactPipelineInfoFolder);
+            }
+            Dictionary<string, HashSet<string>> CIPlan = new Dictionary<string, HashSet<string>>
+            {
+                [BUILD_PHASE] = new HashSet<string>(influencedModuleInfo[BUILD_PHASE].Select(GetModuleNameFromPath).Where(x => x != null)),
+                [ANALYSIS_BREAKING_CHANGE_PHASE] = influencedModuleInfo[ANALYSIS_BREAKING_CHANGE_PHASE],
+                [ANALYSIS_DEPENDENCY_PHASE] = influencedModuleInfo[ANALYSIS_DEPENDENCY_PHASE],
+                [ANALYSIS_HELP_EXAMPLE_PHASE] = influencedModuleInfo[ANALYSIS_HELP_EXAMPLE_PHASE],
+                [ANALYSIS_HELP_PHASE] = influencedModuleInfo[ANALYSIS_HELP_PHASE],
+                [ANALYSIS_SIGNATURE_PHASE] = influencedModuleInfo[ANALYSIS_SIGNATURE_PHASE],
+                [TEST_PHASE] = new HashSet<string>(influencedModuleInfo[TEST_PHASE].Select(GetModuleNameFromPath).Where(x => x != null))
+            };
+            File.WriteAllText(Path.Combine(config.ArtifactPipelineInfoFolder, "CIPlan.json"), JsonConvert.SerializeObject(CIPlan, Formatting.Indented));
+
             return true;
+        }
+
+        private string GetModuleNameFromPath(string path)
+        {
+            if (path.IndexOf("src") == -1)
+            {
+                return null;
+            }
+            return path.Replace("\\", "/").Split("src/")[1].Split('/')[0];
         }
 
         /// <summary>
@@ -431,6 +460,7 @@ namespace Microsoft.WindowsAzure.Build.Tasks
                         [BUILD_PHASE] = new HashSet<string>(selectedModuleList),
                         [ANALYSIS_BREAKING_CHANGE_PHASE] = new HashSet<string>(selectedModuleList),
                         [ANALYSIS_DEPENDENCY_PHASE] = new HashSet<string>(selectedModuleList),
+                        [ANALYSIS_HELP_EXAMPLE_PHASE] = new HashSet<string>(selectedModuleList),
                         [ANALYSIS_HELP_PHASE] = new HashSet<string>(selectedModuleList),
                         [ANALYSIS_SIGNATURE_PHASE] = new HashSet<string>(selectedModuleList),
                         [TEST_PHASE] = new HashSet<string>(selectedModuleList)