+++ src/ApiManagement/ApiManagement/ChangeLog.md
@@ -19,7 +19,7 @@
 -->
 ## Upcoming Release
 
-## Version 3.1.0
+## Version 4.0.0
 * [Breaking Change] Changed the type of parameter `Sku` from Enum to String in `Add-AzApiManagementRegion`, `New-AzApiManagement` and `Update-AzApiManagementRegion`.
 
 ## Version 3.0.1