+++ src/CognitiveServices/CognitiveServices/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.12.0
+* Updated CognitiveServices PowerShell to use 2022-10-01 version.
+
 ## Version 1.11.0
 * Updated CognitiveServices PowerShell to use 2022-03-01 version.
 * Added `Get-AzCognitiveServicesAccountModel` cmdlet.