+++ src/Migrate/Changelog.md
@@ -19,7 +19,11 @@
 -->
 ## Upcoming Release
 
-## Version 1.2.0
+## Version 2.1.0
+* Added parameter `CacheStorageAccountId` to `Initialize-AzMigrateReplicationInfrastructure`
+* Added support for OS Disk Swap and Test Migrate Subnet Selection
+
+## Version 2.0.0
 * Updated ApiVersion to 2022-05-01
 * Added support for pause and resume
   * `Suspend-AzMigrateServerReplication`