+++ src/ServiceBus/ServiceBus/ChangeLog.md
@@ -20,7 +20,7 @@
 -->
 ## Upcoming Release
 
-## Version 1.12.0
+## Version 2.0.0
 * Most cmdlets in Az.ServiceBus module have been migrated to a new format and would witness breaking changes. Please refer our migration guide https://go.microsoft.com/fwlink/?linkid=2204584 to know breaking changes in detail.
 
 ## Version 1.11.0