+++ src/Websites/Websites/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 ## Upcoming Release
 
+## Version 2.11.5
+* Fixed `Publish-AzWebApp` to use latest publish API when deploying war package [#19791]
 ## Version 2.11.4
 * Fixed `Import-AzWebAppKeyVaultCertificate` to use certificate naming convention same as Az-CLI